requirements.txt
streamlit
requests
openapi
python-dotenv
google-generativeai